Big refactoring, small bugfixes, new theme
Important Refactoring : We now truing to respect a little more the mvc
pattern. Better separation between engine, controller and ui. Views
are more passive and queries the controller instead of storing some
part of the data and reading the rest directly in the model. Some bugs
related to this refactoring may remains ...

Imported Aero-ion3.1 gtk theme into the project. This is the new gtk
theme of kiara

EventMerger wasn't correclty the midi chan of its busses. It is now
fixed, whe should be able to play midi on all channels.

Removed the little support we had in playlist and pianoo roll of mouse
events.